refactor: 예상하지 못한 예외에 대한 핸들링 메시지를 프로필에 따라 분리
