+++ backend/src/main/java/com/yigongil/backend/ui/exceptionhandler/InternalServerErrorMessageConverter.java
@@ -1,0 +1,6 @@
+package com.yigongil.backend.ui.exceptionhandler;
+
+public interface InternalServerErrorMessageConverter {
+
+    String convert(Exception e);
+}
